+++ app/src/main/res/layout/edit_player_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditPlayerFragment">
+    tools:context=".fragment.EditPlayerFragment">
 
     <TextView
         android:id="@+id/label_name"
